Moving Position to dom folder
